TruthCheck-Fake-News-Detector-main/Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
numpy>=1.23.2,<2.0
scikit-learn==1.7.1
scipy==1.16.1
joblib==1.5.1
threadpoolctl==3.6.0
Pillow==10.1.0
pytesseract==0.3.10
opencv-python==4.8.1.78